Declare `DummyStoreConfig` in a header
This will useful for unit tests.
